fixed up LOD node selection in lodmap

+++ src/modules/lodmap/lodmap.h
@@ -56,7 +56,7 @@ typedef struct {
 	/** The current selection of quadtree nodes to be rendered */
 	GList *selection;
 	/** The base viewing range covered by the lowest LOD level in the LOD map */
-	unsigned int baseRange;
+	float baseRange;
 	/** The maximum viewing distance to be handled by this LOD map */
 	unsigned int viewingDistance;
 	/** The data prefix to prepend to file names on map loading */
@@ -65,7 +65,7 @@ typedef struct {
 	char *dataSuffix;
 } OpenGLLodMap;
 
-API OpenGLLodMap *createOpenGLLodMap(unsigned int baseRange, unsigned int viewingDistance, unsigned int leafSize, const char *dataPrefix, const char *dataSuffix);
+API OpenGLLodMap *createOpenGLLodMap(double baseRange, unsigned int viewingDistance, unsigned int leafSize, const char *dataPrefix, const char *dataSuffix);
 API void updateOpenGLLodMap(OpenGLLodMap *lodmap, Vector *position);
 API void drawOpenGLLodMap(OpenGLLodMap *lodmap);
 API void freeOpenGLLodMap(OpenGLLodMap *lodmap);
